Moved almost all code behinds to respective ViewModels
