+++ src/Files.App/Data/EventArguments/DrivesWidgetInvokedEventArgs.cs
@@ -1,0 +1,10 @@
+﻿// Copyright (c) 2023 Files Community
+// Licensed under the MIT License. See the LICENSE.
+
+namespace Files.App.Data.EventArguments
+{
+	public class DrivesWidgetInvokedEventArgs : EventArgs
+	{
+		public string? Path { get; set; }
+	}
+}
