+++ src/Files.App/Data/Items/WidgetFileTagCardItem.cs
@@ -9,6 +9,10 @@ namespace Files.App.Data.Items
 {
 	public sealed partial class WidgetFileTagCardItem : WidgetCardItem
 	{
+		// Dependency injections
+
+		public IContentPageContext ContentPageContext { get; } = Ioc.Default.GetRequiredService<IContentPageContext>();
+
 		// Fields
 
 		private readonly IStorable _associatedStorable;
@@ -47,10 +51,9 @@ namespace Files.App.Data.Items
 
 		public ICommand ClickCommand { get; }
 
-		public WidgetFileTagCardItem(IStorable associatedStorable, Func<string, Task> openAction, IImage? icon)
+		public WidgetFileTagCardItem(IStorable associatedStorable, IImage? icon)
 		{
 			_associatedStorable = associatedStorable;
-			_openAction = openAction;
 			_Icon = icon;
 			_Name = associatedStorable.Name;
 			_Path = associatedStorable.TryGetPath();
@@ -61,7 +64,7 @@ namespace Files.App.Data.Items
 
 		private Task ClickAsync(CancellationToken cancellationToken)
 		{
-			return _openAction(_associatedStorable.Id);
+			return NavigationHelpers.OpenPath(_associatedStorable.Id, ContentPageContext.ShellPage!);
 		}
 	}
 }
