+++ src/Files.App/UserControls/Widgets/FileTagsWidget.xaml.cs
@@ -1,14 +1,8 @@
 ﻿// Copyright (c) 2023 Files Community
 // Licensed under the MIT License. See the LICENSE.
 
-using Files.App.Helpers.ContextFlyouts;
-using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
-using Microsoft.UI.Xaml.Controls.Primitives;
 using Microsoft.UI.Xaml.Input;
-using System.IO;
-using System.Windows.Input;
-using Windows.Storage;
 
 namespace Files.App.UserControls.Widgets
 {
@@ -16,255 +10,24 @@ namespace Files.App.UserControls.Widgets
 	/// Represents group of control displays a list of <see cref="WidgetFileTagsContainerItem"/>
 	/// and its inner items with <see cref="WidgetFileTagCardItem"/>.
 	/// </summary>
-	public sealed partial class FileTagsWidget : BaseWidgetViewModel, IWidgetViewModel
+	public sealed partial class FileTagsWidget : UserControl
 	{
-		private IHomePageContext HomePageContext { get; } = Ioc.Default.GetRequiredService<IHomePageContext>();
-
-		public FileTagsWidgetViewModel ViewModel
-		{
-			get => (FileTagsWidgetViewModel)DataContext;
-			set => DataContext = value;
-		}
-
-		public IShellPage AppInstance;
-
-		public Func<string, Task>? OpenAction { get; set; }
-
-		public delegate void FileTagsOpenLocationInvokedEventHandler(object sender, PathNavigationEventArgs e);
-		public delegate void FileTagsNewPaneInvokedEventHandler(object sender, QuickAccessCardInvokedEventArgs e);
-		public static event EventHandler<IEnumerable<WidgetFileTagCardItem>>? SelectedTaggedItemsChanged;
-		public event FileTagsOpenLocationInvokedEventHandler FileTagsOpenLocationInvoked;
-		public event FileTagsNewPaneInvokedEventHandler FileTagsNewPaneInvoked;
-
-		public string WidgetName => nameof(FileTagsWidget);
-		public string WidgetHeader => "FileTags".GetLocalizedResource();
-		public string AutomationProperties => "FileTags".GetLocalizedResource();
-		public bool IsWidgetSettingEnabled => UserSettingsService.GeneralSettingsService.ShowFileTagsWidget;
-		public bool ShowMenuFlyout => false;
-		public MenuFlyoutItem? MenuFlyoutItem => null;
-
-		private ICommand OpenInNewPaneCommand;
+		public FileTagsWidgetViewModel ViewModel { get; } = new();
 
 		public FileTagsWidget()
 		{
 			InitializeComponent();
-
-			// Second function is layered on top to ensure that OpenPath function is late initialized and a null reference is not passed-in
-			// See FileTagItemViewModel._openAction for more information
-			ViewModel = new(x => OpenAction!(x));
-			OpenInNewTabCommand = new AsyncRelayCommand<WidgetCardItem>(OpenInNewTabAsync);
-			OpenInNewWindowCommand = new AsyncRelayCommand<WidgetCardItem>(OpenInNewWindowAsync);
-			OpenFileLocationCommand = new RelayCommand<WidgetCardItem>(OpenFileLocation);
-			OpenInNewPaneCommand = new RelayCommand<WidgetCardItem>(OpenInNewPane);
-			PinToFavoritesCommand = new AsyncRelayCommand<WidgetCardItem>(PinToFavoritesAsync);
-			UnpinFromFavoritesCommand = new AsyncRelayCommand<WidgetCardItem>(UnpinFromFavoritesAsync);
-			OpenPropertiesCommand = new RelayCommand<WidgetCardItem>(OpenProperties);
-		}
-
-		private void OpenProperties(WidgetCardItem? item)
-		{
-			if (!HomePageContext.IsAnyItemRightClicked)
-				return;
-
-			var flyout = HomePageContext.ItemContextFlyoutMenu;
-			EventHandler<object> flyoutClosed = null!;
-			flyoutClosed = (s, e) =>
-			{
-				flyout!.Closed -= flyoutClosed;
-
-				ListedItem listedItem = new(null!)
-				{
-					ItemPath = (item.Item as WidgetFileTagCardItem)?.Path ?? string.Empty,
-					ItemNameRaw = (item.Item as WidgetFileTagCardItem)?.Name ?? string.Empty,
-					PrimaryItemAttribute = StorageItemTypes.Folder,
-					ItemType = "Folder".GetLocalizedResource(),
-				};
-				FilePropertiesHelpers.OpenPropertiesWindow(listedItem, AppInstance);
-			};
-
-			flyout!.Closed += flyoutClosed;
-		}
-
-		private void OpenInNewPane(WidgetCardItem? item)
-		{
-			FileTagsNewPaneInvoked?.Invoke(this, new QuickAccessCardInvokedEventArgs()
-			{
-				Path = item?.Path ?? string.Empty
-			});
-		}
-
-		private async void FileTagItem_ItemClick(object sender, ItemClickEventArgs e)
-		{
-			if (e.ClickedItem is WidgetFileTagCardItem itemViewModel)
-				itemViewModel.ClickCommand.Execute(null);
 		}
 
 		private void AdaptiveGridView_RightTapped(object sender, RightTappedRoutedEventArgs e)
 		{
-			// Ensure values are not null
-			if (e.OriginalSource is not FrameworkElement element ||
-				element.DataContext is not WidgetFileTagCardItem item)
-				return;
-
-			// Create a new Flyout
-			var itemContextMenuFlyout = new CommandBarFlyout()
-			{
-				Placement = FlyoutPlacementMode.Full
-			};
-
-			// Hook events
-			itemContextMenuFlyout.Opening += (sender, e) => App.LastOpenedFlyout = sender as CommandBarFlyout;
-			itemContextMenuFlyout.Closed += (sender, e) => OnRightClickedItemChanged(null, null);
-
-			_flyoutItemPath = item.Path;
-
-			// Notify of the change on right clicked item
-			OnRightClickedItemChanged(item, itemContextMenuFlyout);
-
-			// Get items for the flyout
-			var menuItems = GetItemMenuItems(item, QuickAccessService.IsItemPinned(item.Path), item.IsFolder);
-			var (_, secondaryElements) = ContextFlyoutModelToElementHelper.GetAppBarItemsFromModel(menuItems);
-
-			// Set max width of the flyout
-			secondaryElements
-				.OfType<FrameworkElement>()
-				.ForEach(i => i.MinWidth = Constants.UI.ContextMenuItemsMaxWidth);
-
-			// Add menu items to the secondary flyout
-			secondaryElements.ForEach(itemContextMenuFlyout.SecondaryCommands.Add);
-
-			// Show the flyout
-			itemContextMenuFlyout.ShowAt(element, new() { Position = e.GetPosition(element) });
-
-			// Load shell menu items
-			_ = ShellContextFlyoutFactory.LoadShellMenuItemsAsync(_flyoutItemPath, itemContextMenuFlyout);
-
-			e.Handled = true;
+			ViewModel.BuildContextFlyout(sender, e);
 		}
 
-		public override List<ContextMenuFlyoutItemViewModel> GetItemMenuItems(WidgetCardItem item, bool isPinned, bool isFolder = false)
+		private void FileTagItem_ItemClick(object sender, ItemClickEventArgs e)
 		{
-			return new List<ContextMenuFlyoutItemViewModel>()
-			{
-				new ContextMenuFlyoutItemViewModel()
-				{
-					Text = "OpenWith".GetLocalizedResource(),
-					OpacityIcon = new OpacityIconModel()
-					{
-						OpacityIconStyle = "ColorIconOpenWith",
-					},
-					Tag = "OpenWithPlaceholder",
-					ShowItem = !isFolder
-				},
-				new ContextMenuFlyoutItemViewModel()
-				{
-					Text = "SendTo".GetLocalizedResource(),
-					Tag = "SendToPlaceholder",
-					ShowItem = !isFolder && UserSettingsService.GeneralSettingsService.ShowSendToMenu
-				},
-				new ContextMenuFlyoutItemViewModel()
-				{
-					Text = "OpenInNewTab".GetLocalizedResource(),
-					OpacityIcon = new OpacityIconModel()
-					{
-						OpacityIconStyle = "ColorIconOpenInNewTab",
-					},
-					Command = OpenInNewTabCommand,
-					CommandParameter = item,
-					ShowItem = isFolder
-				},
-				new ContextMenuFlyoutItemViewModel()
-				{
-					Text = "OpenInNewWindow".GetLocalizedResource(),
-					OpacityIcon = new OpacityIconModel()
-					{
-						OpacityIconStyle = "ColorIconOpenInNewWindow",
-					},
-					Command = OpenInNewWindowCommand,
-					CommandParameter = item,
-					ShowItem = isFolder
-				},
-				new ContextMenuFlyoutItemViewModel()
-				{
-					Text = "OpenFileLocation".GetLocalizedResource(),
-					Glyph = "\uED25",
-					Command = OpenFileLocationCommand,
-					CommandParameter = item,
-					ShowItem = !isFolder
-				},
-				new ContextMenuFlyoutItemViewModel()
-				{
-					Text = "OpenInNewPane".GetLocalizedResource(),
-					Command = OpenInNewPaneCommand,
-					CommandParameter = item,
-					ShowItem = UserSettingsService.GeneralSettingsService.ShowOpenInNewPane && isFolder
-				},
-				new ContextMenuFlyoutItemViewModel()
-				{
-					Text = "PinToFavorites".GetLocalizedResource(),
-					OpacityIcon = new OpacityIconModel()
-					{
-						OpacityIconStyle = "ColorIconPinToFavorites",
-					},
-					Command = PinToFavoritesCommand,
-					CommandParameter = item,
-					ShowItem = !isPinned && isFolder
-				},
-				new ContextMenuFlyoutItemViewModel()
-				{
-					Text = "UnpinFromFavorites".GetLocalizedResource(),
-					OpacityIcon = new OpacityIconModel()
-					{
-						OpacityIconStyle = "ColorIconUnpinFromFavorites",
-					},
-					Command = UnpinFromFavoritesCommand,
-					CommandParameter = item,
-					ShowItem = isPinned && isFolder
-				},
-				new ContextMenuFlyoutItemViewModel()
-				{
-					Text = "Properties".GetLocalizedResource(),
-					OpacityIcon = new OpacityIconModel()
-					{
-						OpacityIconStyle = "ColorIconProperties",
-					},
-					Command = OpenPropertiesCommand,
-					CommandParameter = item,
-					ShowItem = isFolder
-				},
-				new ContextMenuFlyoutItemViewModel()
-				{
-					ItemType = ContextMenuFlyoutItemType.Separator,
-					Tag = "OverflowSeparator",
-				},
-				new ContextMenuFlyoutItemViewModel()
-				{
-					Text = "Loading".GetLocalizedResource(),
-					Glyph = "\xE712",
-					Items = new List<ContextMenuFlyoutItemViewModel>(),
-					ID = "ItemOverflow",
-					Tag = "ItemOverflow",
-					IsEnabled = false,
-				}
-			}.Where(x => x.ShowItem).ToList();
-		}
-
-		public void OpenFileLocation(WidgetCardItem? item)
-		{
-			FileTagsOpenLocationInvoked?.Invoke(this, new PathNavigationEventArgs()
-			{
-				ItemPath = Directory.GetParent(item?.Path ?? string.Empty)?.FullName ?? string.Empty,
-				ItemName = Path.GetFileName(item?.Path ?? string.Empty),
-			});
-		}
-
-		public Task RefreshWidgetAsync()
-		{
-			return Task.CompletedTask;
-		}
-
-		public void Dispose()
-		{
+			if (e.ClickedItem is WidgetFileTagCardItem itemViewModel)
+				itemViewModel.ClickCommand.Execute(null);
 		}
 	}
 }
