+++ src/Files.App/UserControls/Widgets/QuickAccessWidget.xaml
@@ -1,5 +1,5 @@
 ﻿<!--  Copyright (c) 2023 Files Community. Licensed under the MIT License. See the LICENSE.  -->
-<local:BaseWidgetViewModel
+<UserControl
 	x:Class="Files.App.UserControls.Widgets.QuickAccessWidget"
 	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
 	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
@@ -7,82 +7,75 @@
 	xmlns:dataitems="using:Files.App.Data.Items"
 	xmlns:local="using:Files.App.UserControls.Widgets"
 	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+	DataContext="{x:Bind ViewModel, Mode=OneWay}"
 	mc:Ignorable="d">
 
-	<Grid>
-		<ItemsRepeater
-			x:Name="CardsList"
-			HorizontalAlignment="Stretch"
-			ItemsSource="{x:Bind local:QuickAccessWidget.ItemsAdded, Mode=OneWay}">
+	<ItemsRepeater ItemsSource="{x:Bind local:QuickAccessWidget.ItemsAdded, Mode=OneWay}">
 
-			<ItemsRepeater.Layout>
-				<UniformGridLayout
-					ItemsStretch="Fill"
-					MaximumRowsOrColumns="6"
-					MinColumnSpacing="8"
-					MinItemHeight="72"
-					MinItemWidth="100"
-					MinRowSpacing="8"
-					Orientation="Horizontal" />
-			</ItemsRepeater.Layout>
+		<ItemsRepeater.ItemTemplate>
+			<DataTemplate x:DataType="dataitems:WidgetFolderCardItem">
+				<Button
+					Padding="8"
+					HorizontalAlignment="Stretch"
+					VerticalAlignment="Stretch"
+					HorizontalContentAlignment="Stretch"
+					VerticalContentAlignment="Stretch"
+					AutomationProperties.Name="{x:Bind AutomationProperties}"
+					Click="Button_Click"
+					CornerRadius="{StaticResource ControlCornerRadius}"
+					DataContext="{x:Bind}"
+					PointerPressed="Button_PointerPressed"
+					RightTapped="Button_RightTapped"
+					Tag="{x:Bind Path}">
+					<Grid>
+						<Grid.RowDefinitions>
+							<RowDefinition Height="*" />
+							<RowDefinition Height="Auto" />
+						</Grid.RowDefinitions>
 
-			<ItemsRepeater.ItemTemplate>
-				<DataTemplate x:DataType="dataitems:WidgetFolderCardItem">
-					<Button
-						Padding="8"
-						HorizontalAlignment="Stretch"
-						VerticalAlignment="Stretch"
-						HorizontalContentAlignment="Stretch"
-						VerticalContentAlignment="Stretch"
-						AutomationProperties.Name="{x:Bind AutomationProperties}"
-						Click="Button_Click"
-						CornerRadius="{StaticResource ControlCornerRadius}"
-						DataContext="{x:Bind}"
-						PointerPressed="Button_PointerPressed"
-						RightTapped="Button_RightTapped"
-						Tag="{x:Bind Path}">
-						<Grid
-							HorizontalAlignment="Stretch"
-							VerticalAlignment="Stretch"
-							RowSpacing="0">
-							<Grid.RowDefinitions>
-								<RowDefinition Height="*" />
-								<RowDefinition Height="Auto" />
-							</Grid.RowDefinitions>
+						<FontIcon
+							x:Name="PinIcon"
+							HorizontalAlignment="Right"
+							VerticalAlignment="Top"
+							x:Load="{x:Bind IsPinned, Mode=OneWay}"
+							FontSize="12"
+							Glyph="&#xE840;" />
 
-							<FontIcon
-								x:Name="PinIcon"
-								HorizontalAlignment="Right"
-								VerticalAlignment="Top"
-								x:Load="{x:Bind IsPinned, Mode=OneWay}"
-								FontSize="12"
-								Glyph="&#xE840;" />
+						<Image
+							Grid.Row="0"
+							Width="32"
+							Height="32"
+							HorizontalAlignment="Center"
+							VerticalAlignment="Center"
+							Source="{x:Bind Thumbnail, Mode=OneWay}"
+							Stretch="Uniform" />
 
-							<Image
-								Grid.Row="0"
-								Width="32"
-								Height="32"
-								HorizontalAlignment="Center"
-								VerticalAlignment="Center"
-								Source="{x:Bind Thumbnail, Mode=OneWay}"
-								Stretch="Uniform" />
+						<TextBlock
+							x:Name="ItemLocationName"
+							Grid.Row="1"
+							HorizontalAlignment="Center"
+							VerticalAlignment="Center"
+							FontSize="14"
+							HorizontalTextAlignment="Center"
+							Text="{x:Bind Text}"
+							TextTrimming="WordEllipsis"
+							TextWrapping="NoWrap" />
 
-							<TextBlock
-								x:Name="ItemLocationName"
-								Grid.Row="1"
-								HorizontalAlignment="Center"
-								VerticalAlignment="Center"
-								FontSize="14"
-								HorizontalTextAlignment="Center"
-								Text="{x:Bind Text}"
-								TextTrimming="WordEllipsis"
-								TextWrapping="NoWrap" />
+					</Grid>
+				</Button>
+			</DataTemplate>
+		</ItemsRepeater.ItemTemplate>
 
-						</Grid>
-					</Button>
-				</DataTemplate>
-			</ItemsRepeater.ItemTemplate>
+		<ItemsRepeater.Layout>
+			<UniformGridLayout
+				ItemsStretch="Fill"
+				MaximumRowsOrColumns="6"
+				MinColumnSpacing="8"
+				MinItemHeight="72"
+				MinItemWidth="100"
+				MinRowSpacing="8"
+				Orientation="Horizontal" />
+		</ItemsRepeater.Layout>
 
-		</ItemsRepeater>
-	</Grid>
-</local:BaseWidgetViewModel>
+	</ItemsRepeater>
+</UserControl>
